Removing "selectable" field from Edge API docs
I don't know if this field was deprecated or removed or other. When I
try to use this field in my project, there is no "selectable" field on a
`FlowEdge`.

The codebase also mentions "selectable" in a few other places, so if
this change is valid, I don't know if other places need to be updated as
well.